// synapse app/types.ts
export interface Step {
    type: 'exercise' | 'rest';
    name: string;
    details: string; // e.g., "3x 12 repetições" or "45 segundos"
}

export interface ActivityItem {
    id: string;
    category: 'workout' | 'meditation' | 'nutrition';
    title: string;
    description: string;
    duration: string;
    Icon: React.FC<React.SVGProps<SVGSVGElement>>;
    bgColor: string;
    textColor: string;
    imageUrl?: string;
    videoUrl?: string;
    audioUrl?: string;
    content: string;
    difficulty?: 'Iniciante' | 'Médio' | 'Difícil';
    goal?: 'Condicionamento' | 'Emagrecimento' | 'Hipertrofia';
    muscleGroup?: string;
    steps?: Step[];
}